requirements.txt
python-socketio<5.0.0
requests>=2.31.0
flask>=3.0.3
waitress>=3.0.0